pyproject.toml
[project]
name = "adaptive-rebalancing-agent"
version = "0.2.0"
description = "Adaptive Portfolio Rebalancing Agent with Ollama + LangGraph"
authors = [
    { name = "Carlo" }
]
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    "streamlit>=1.40",
    "pandas>=2.2",
    "numpy>=1.26",
    "yfinance>=0.2",
    "pyportfolioopt>=1.5",
    "ollama>=0.3.0",
    "pyyaml>=6.0",
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",
]